Add modulus/multichoose_simple.hpp

+++ modulus/choose_simple.hpp
@@ -7,7 +7,7 @@
  * @brief combination / 組合せ ${} _ n C _ r$ (愚直 $O(r)$)
  */
 template <int32_t MOD>
-mint<MOD> choose_simple(int n, int r) {
+mint<MOD> choose_simple(int64_t n, int32_t r) {
     assert (0 <= r and r <= n);
     mint<MOD> num = 1;
     mint<MOD> den = 1;
